moved js-joda date method to DateUtil service
OCD-3456
